Combine Excel and deployment credential encryption in one file
